okay

+++ GroceryWebapp/main.py
@@ -3,7 +3,7 @@ from . import db
 from .models import User,Categories,Products,Cart
 from datetime import datetime
 from flask_sqlalchemy import SQLAlchemy
-from flask_login import LoginManager, UserMixin, login_user, logout_user, login_required, current_user
+from flask_login import login_required, current_user
 
 
 global_user=None
@@ -58,7 +58,7 @@ def adminhome(bool1,userid):
 @login_required
 def adminhomecategories():
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         
         if request.method == 'POST':
             cat_name = request.form['category_name']
@@ -71,6 +71,8 @@ def adminhomecategories():
                 db.session.commit()
         data = Categories.query.all()
         return render_template("home_admin.html", bool=True, data=data)
+    else:
+        return("<center><h1>Access Denied</h1></center>")    
 
 
 # Route to display all categories in the admin home page
@@ -78,7 +80,7 @@ def adminhomecategories():
 @login_required
 def adminshowcategories():
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         data = Categories.query.all()
         return render_template("home_admin.html", bool=True, data=data, boolcat=True)
 
@@ -88,7 +90,7 @@ def adminshowcategories():
 @login_required
 def update_category(categories_id):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
     # Retrieve the category object based on the provided ID
         ncat = Categories.query.get_or_404(categories_id)
         if request.method == "POST":
@@ -99,14 +101,15 @@ def update_category(categories_id):
             return redirect(url_for("main.adminhome", bool1=False))
 
         return render_template("update_categories.html", categories=ncat)
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to add a new item to a category
 @main.route("/additem/<string:categories_id>", methods=["GET", "POST"])
 @login_required
 def add_item(categories_id):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         if request.method == "POST":
             # Retrieve the product details from the form submission
             pname = request.form.get("product_name")
@@ -128,13 +131,14 @@ def add_item(categories_id):
             db.session.commit()
 
         return render_template("additem.html", cid=categories_id)
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 @main.route("/viewitems/<string:categories_id>/<string:type>", methods=["GET", "POST"])
 @login_required
 def view_items(categories_id, type):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         data = Products.query.filter_by(c_id=categories_id).all()
         return render_template("views_items.html", prod=data, check=type)
 
@@ -144,7 +148,7 @@ def view_items(categories_id, type):
 @login_required
 def update_item(item_id, type):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         # Retrieve the item object based on the provided ID
         nitem = Products.query.get_or_404(item_id)
         if request.method == "POST":
@@ -161,26 +165,28 @@ def update_item(item_id, type):
             return render_template("views_items.html", prod=data, check="1")
 
         return render_template("views_items.html", item_id=item_id, check=type)
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to revert changes to an item
 @main.route("/revertitem/<string:item_id>/<string:type>", methods=["GET", "POST"])
 @login_required
 def revert_item(item_id, type):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         nitem = Products.query.get_or_404(item_id)
         categories_id = nitem.c_id
         data = Products.query.filter_by(c_id=categories_id).all()
         return render_template("views_items.html", prod=data, check=type)
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to delete an item
 @main.route("/delitem/<string:item_id>", methods=["GET", "POST"])
 @login_required
 def delete_item(item_id):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         nitem = Products.query.get_or_404(item_id)
         db.session.delete(nitem)
         db.session.commit()
@@ -188,14 +194,15 @@ def delete_item(item_id):
         categories_id = nitem.c_id
         data = Products.query.filter_by(c_id=categories_id).all()
         return render_template("views_items.html", prod=data, check="1")
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to delete a category and all associated items
 @main.route("/deletecategory/<string:categories_id>", methods=["GET", "POST"])
 @login_required
 def delete_category(categories_id):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         try:
             # Delete associated products first
             data = Products.query.filter_by(c_id=categories_id).all()
@@ -211,7 +218,8 @@ def delete_category(categories_id):
             db.session.rollback()  
             datacat = Categories.query.all()
             return render_template("home_admin.html", bool=True, data=datacat, boolcat=True)
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 
 # Route to add an item to the cart
@@ -219,7 +227,7 @@ def delete_category(categories_id):
 @login_required
 def addtocart(product_id, userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         # Retrieve the product based on the provided ID
         product_id = int(product_id)
         product = Products.query.get(product_id)
@@ -253,7 +261,8 @@ def addtocart(product_id, userid):
             total_price=total_price,
             userid=userid,  
         )
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to save the cart after adding items
 @main.route(
@@ -263,7 +272,7 @@ def addtocart(product_id, userid):
 @login_required
 def savecart(productid, quantity, total, userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         # Retrieve the product based on the provided ID
         items = Products.query.get_or_404(productid)
         categoryid = items.c_id
@@ -302,14 +311,15 @@ def savecart(productid, quantity, total, userid):
             db.session.commit()
 
         return redirect(url_for("main.addtocart", product_id=productid, userid=userid))
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to show the cart items for a user
 @login_required
 @main.route("/showcart/<string:productid>/<int:userid>", methods=["GET", "POST"])
 def showcart(productid, userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         cartdata = Cart.query.filter_by(cart_user_id=userid).all()
         user = User.query.get_or_404(userid)
         items = Products.query.get_or_404(productid)
@@ -321,6 +331,8 @@ def showcart(productid, userid):
         return render_template(
             "cart.html", cartdata=cartdata, user=user, totalprice=totalprice
         )
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 
 # Route to delete an item from the cart
@@ -328,7 +340,7 @@ def showcart(productid, userid):
 @main.route("/delcart/<int:cartid>/<int:userid>", methods=["GET", "POST"])
 def delete_cart(cartid, userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
 
         ncart = Cart.query.get_or_404(cartid)
         restock = ncart.cart_quantity
@@ -349,14 +361,15 @@ def delete_cart(cartid, userid):
         return render_template(
             "cart.html", user=user, cartdata=cartdata, totalprice=totalprice
         )
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to view cart items for a user
 @main.route("/usercart/<int:userid>", methods=["GET", "POST"])
 @login_required
 def usercart(userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         cartdata = Cart.query.filter_by(cart_user_id=userid).all()
         user = User.query.get_or_404(userid)
         usercartrecord = Cart.query.filter_by(cart_user_id=userid).all()
@@ -367,36 +380,39 @@ def usercart(userid):
         return render_template(
             "cart.html", cartdata=cartdata, user=user, totalprice=totalprice
         )
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to complete the purchase and clear the cart
 @main.route("/purchase/<int:userid>", methods=["GET", "POST"])
 @login_required
 def purchase(userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         cartdata = Cart.query.filter_by(cart_user_id=userid).all()
         for cart in cartdata:
             db.session.delete(cart)
             db.session.commit()
         user = User.query.get_or_404(userid)
         return render_template("cart.html", user=user, purchbool=True)
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 @main.route("/searchcat/<int:userid>", methods=["GET", "POST"])
 @login_required
 def searchcat(userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         return render_template("Search_elems.html", userid=userid, role="searchcateg")
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to perform a search based on a category name
 @main.route("/search_categ/<int:userid>", methods=["GET", "POST"])
 @login_required
 def search_categ(userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         catname = request.form.get("category_name")
         categories = Categories.query.all()
         for cat in categories:
@@ -411,7 +427,8 @@ def search_categ(userid):
         return render_template(
             "showsearch.html", role="notfound", userid=userid, search="categories"
         )
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to search for items
 @main.route("/searchitem/<int:userid>", methods=["GET", "POST"])
@@ -420,14 +437,15 @@ def searchitem(userid):
     global global_user
     if(current_user.id==global_user):
         return render_template("Search_elems.html", userid=userid, role="searchitems")
-
+    else:
+        return("<center><h1>Access Denied</h1></center>")
 
 # Route to perform a search based on item name, price, manufacture date, and expiry date
 @login_required
 @main.route("/search_item/<int:userid>", methods=["GET", "POST"])
 def search_item(userid):
     global global_user
-    if(current_user.id==global_user):
+    if current_user.is_authenticated and current_user.id == global_user:
         itemname = request.form.get("product_name")
         itemprice = request.form.get("price")
         if itemprice:
@@ -479,3 +497,5 @@ def search_item(userid):
         return render_template(
             "showsearch.html", userid=userid, search="items", role="notfound"
         )
+    else:
+        return("<center><h1>Access Denied</h1></center>")
